fix(publikator): extract script/lib/transformPublications
